+++ test/hui_test.exs
@@ -128,42 +128,79 @@ defmodule HuiTest do
       assert %Hui.F.Range{
         per_field: false,
         range: "year",
-        "range.end": 1799,
-        "range.gap": "+10YEARS",
-        "range.hardend": nil,
-        "range.include": nil,
-        "range.method": nil,
-        "range.other": nil,
-        "range.start": 1700
-      } = %Hui.F.Range{range: "year", "range.gap": "+10YEARS", "range.start": 1700, "range.end": 1799}
+        end: 1799,
+        gap: "+10YEARS",
+        hardend: nil,
+        include: nil,
+        method: nil,
+        other: nil,
+        start: 1700
+      } = %Hui.F.Range{range: "year", gap: "+10YEARS", start: 1700, end: 1799}
     end
 
     test "interval parameter setting" do
       assert %Hui.F.Interval{
         interval: "price",
-        "interval.set": ["[0,10]", "(10,100]"],
+        set: ["[0,10]", "(10,100]"],
         per_field: false
-      } = %Hui.F.Interval{interval: "price", "interval.set": ["[0,10]", "(10,100]"]}
+      } = %Hui.F.Interval{interval: "price", set: ["[0,10]", "(10,100]"]}
     end
 
     test "can be encoded into URL query string" do
-      assert "facet=true&facet.field=type&facet.field=year&facet.query=year%3A%5B2000+TO+NOW%5D" = %Hui.F{field: ["type", "year"], query: "year:[2000 TO NOW]"} |> Hui.URL.encode_query
+      assert "facet=true&facet.field=type&facet.field=year&facet.query=year%3A%5B2000+TO+NOW%5D" 
+      = %Hui.F{field: ["type", "year"], query: "year:[2000 TO NOW]"} |> Hui.URL.encode_query
     end
 
     test "range can be encoded into URL query string" do
-      x = %Hui.F.Range{range: "year", "range.gap": "+10YEARS", "range.start": 1700, "range.end": 1799}
-      assert "facet.range=year&facet.range.end=1799&facet.range.gap=%2B10YEARS&facet.range.start=1700" = x |> Hui.URL.encode_query
+      x = %Hui.F.Range{range: "year", gap: "+10YEARS", start: 1700, end: 1799}
+      assert "facet.range.end=1799&facet.range.gap=%2B10YEARS&facet.range=year&facet.range.start=1700"
+      = x |> Hui.URL.encode_query
 
       y = %Hui.F{field: "type", range: x}
-      assert "facet=true&facet.field=type&facet.range=year&facet.range.end=1799&facet.range.gap=%2B10YEARS&facet.range.start=1700" = y |> Hui.URL.encode_query
+      assert "facet=true&facet.field=type&facet.range.end=1799&facet.range.gap=%2B10YEARS&facet.range=year&facet.range.start=1700" 
+      = y |> Hui.URL.encode_query
     end
 
     test "interval can be encoded into URL query string" do
-      x = %Hui.F.Interval{interval: "price", "interval.set": ["[0,10]", "(10,100]"]}
-      assert "facet.interval=price&facet.interval.set=%5B0%2C10%5D&facet.interval.set=%2810%2C100%5D" = x |> Hui.URL.encode_query
+      x = %Hui.F.Interval{interval: "price", set: ["[0,10]", "(10,100]"]}
+      assert "facet.interval=price&facet.interval.set=%5B0%2C10%5D&facet.interval.set=%2810%2C100%5D"
+      = x |> Hui.URL.encode_query
 
       y = %Hui.F{field: "type", interval: x}
-      assert "facet=true&facet.field=type&facet.interval=price&facet.interval.set=%5B0%2C10%5D&facet.interval.set=%2810%2C100%5D" = y |> Hui.URL.encode_query
+      assert "facet=true&facet.field=type&facet.interval=price&facet.interval.set=%5B0%2C10%5D&facet.interval.set=%2810%2C100%5D"
+      = y |> Hui.URL.encode_query
+    end
+
+    test "range per field URL encoding" do
+     x = %Hui.F.Range{range: "year", gap: "+10YEARS", start: 1700, end: 1799, per_field: true}
+     assert "f.year.facet.range.end=1799&f.year.facet.range.gap=%2B10YEARS&facet.range=year&f.year.facet.range.start=1700"
+     = x |> Hui.URL.encode_query
+
+     y = %Hui.F.Range{range: "price", gap: "10", start: 0, end: 100, per_field: true}
+     assert "f.price.facet.range.end=100&f.price.facet.range.gap=10&facet.range=price&f.price.facet.range.start=0"
+     = y |> Hui.URL.encode_query
+
+     z = %Hui.F{field: "type", range: [x, y]}
+     assert "facet=true&facet.field=type&" <>
+     "f.year.facet.range.end=1799&f.year.facet.range.gap=%2B10YEARS&facet.range=year&f.year.facet.range.start=1700&" <>
+     "f.price.facet.range.end=100&f.price.facet.range.gap=10&facet.range=price&f.price.facet.range.start=0"
+     = z |> Hui.URL.encode_query     
+    end
+
+    test "interval per field URL encoding" do
+     x = %Hui.F.Interval{interval: "price", set: ["[0,10]", "(10,100]"], per_field: true}
+     assert "facet.interval=price&f.price.facet.interval.set=%5B0%2C10%5D&f.price.facet.interval.set=%2810%2C100%5D"
+     = x |> Hui.URL.encode_query
+
+     y = %Hui.F.Interval{interval: "age", set: ["[0,30]", "(30,60]", "[60, 100]"], per_field: true}
+     assert "facet.interval=age&f.age.facet.interval.set=%5B0%2C30%5D&f.age.facet.interval.set=%2830%2C60%5D&f.age.facet.interval.set=%5B60%2C+100%5D"
+     = y |> Hui.URL.encode_query
+
+     z = %Hui.F{field: "type", interval: [x, y]}
+     assert "facet=true&facet.field=type&" <>
+     "facet.interval=price&f.price.facet.interval.set=%5B0%2C10%5D&f.price.facet.interval.set=%2810%2C100%5D&" <>
+     "facet.interval=age&f.age.facet.interval.set=%5B0%2C30%5D&f.age.facet.interval.set=%2830%2C60%5D&f.age.facet.interval.set=%5B60%2C+100%5D"
+     = z |> Hui.URL.encode_query
     end
 
   end